Move project related methods to separate source file
